fix(baseobject): COLLISION msg carries full pointer, not truncated int32
DispatchCollisionMessages passed the colliding object's address as int32 data,
silently truncating on 64-bit. Receivers (enemy.cc, explode.cc) cast the buffer
to Actor* and dereferenced — reading garbage on arm64. On 32-bit the truncation
was a no-op so the bug was latent; arm64 surfaced it as a compiler error.

Fix: widen SMsg::_data._message to uintptr_t so the union holds a pointer on
any platform. PutMsg/sendMsg scalar overload take uintptr_t; existing integer
callers (DELTA_HEALTH etc.) widen implicitly and readers of the low 4 bytes
via *(int32*)msgData still see the original value on little-endian.

Collision sender now uses reinterpret_cast<uintptr_t>(&obj); enemy/explode
receivers read *(uintptr_t*)msgData before casting to Actor*. Receivers that
discard the data (movement.cc, movepath.cc, movefoll.cc, movecam.cc) unchanged.

Plan: docs/plans/2026-04-17-collision-message-pointer-fix.md

+++ wfsource/source/baseobject/baseobject.cc
@@ -56,7 +56,7 @@ operator << ( std::ostream& s, const BaseObject& obj )
 //==============================================================================
                       
 bool
-BaseObject::sendMsg(const int16 msgType, const int32 msgData)
+BaseObject::sendMsg(const int16 msgType, const uintptr_t msgData)
 {
 	bool bSuccess = GetMsgPort().PutMsg( msgType, msgData );
 	assert( bSuccess );
